chore: align quality gate commands
